requirements.txt
fastapi==0.109.0
openai
jinja2
python-multipart
boto3
asyncpg
python-jose[cryptography]
python-dotenv
sqlalchemy
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
typing_extensions
psycopg2-binary
fastapi_sessions
starlette
langchain
openai
pyautogen
jupyterlab
pygithub